Add a naive mutex implementation because unittests don't execute sequentially.
This is super naive and will suffer from race conditions.
But transform() should only be used by one test function at once.
unittest doesn't allow strictly sequential execution of the test cases.
And using asyncio doesn't work out of the box because idk.

So for our purpose it is enough.